requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Environment and configuration
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# CORS and middleware
python-multipart==0.0.6

# Development dependencies (optional, uncomment if needed)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2
